Rename `exception_handler` to `exceptions`, and move deprecation warnings to the renamed `exceptions` module
